Move packaging tests to right file
